Make getElementData shared code
