+++ Server/mods/deathmatch/logic/luadefs/CLuaElementDefs.cpp
@@ -68,7 +68,7 @@ void CLuaElementDefs::LoadFunctions()
         {"getElementAttachedOffsets", getElementAttachedOffsets},
 
         // Element data
-        {"getElementData", getElementData},
+        {"getElementData", GetElementData},
         {"setElementData", setElementData},
         {"removeElementData", removeElementData},
 
@@ -487,46 +487,6 @@ int CLuaElementDefs::getElementByIndex(lua_State* luaVM)
         {
             lua_pushelement(luaVM, pElement);
             return 1;
-        }
-    }
-    else
-        m_pScriptDebugging->LogCustom(luaVM, argStream.GetFullErrorMessage());
-
-    lua_pushboolean(luaVM, false);
-    return 1;
-}
-
-int CLuaElementDefs::getElementData(lua_State* luaVM)
-{
-    //  var getElementData ( element theElement, string key [, inherit = true] )
-    CElement* pElement;
-    SString   strKey;
-    bool      bInherit;
-
-    CScriptArgReader argStream(luaVM);
-    argStream.ReadUserData(pElement);
-    argStream.ReadString(strKey);
-    argStream.ReadBool(bInherit, true);
-
-    if (!argStream.HasErrors())
-    {
-        CLuaMain* pLuaMain = m_pLuaManager->GetVirtualMachine(luaVM);
-        if (pLuaMain)
-        {
-            if (strKey.length() > MAX_CUSTOMDATA_NAME_LENGTH)
-            {
-                // Warn and truncate if key is too long
-                m_pScriptDebugging->LogCustom(luaVM, SString("Truncated argument @ '%s' [%s]", lua_tostring(luaVM, lua_upvalueindex(1)),
-                                                             *SString("string length reduced to %d characters at argument 2", MAX_CUSTOMDATA_NAME_LENGTH)));
-                strKey = strKey.Left(MAX_CUSTOMDATA_NAME_LENGTH);
-            }
-
-            CLuaArgument* pVariable = CStaticFunctionDefinitions::GetElementData(pElement, strKey, bInherit);
-            if (pVariable)
-            {
-                pVariable->Push(luaVM);
-                return 1;
-            }
         }
     }
     else
