+++ Server/mods/deathmatch/logic/luadefs/CLuaElementDefs.h
@@ -67,7 +67,7 @@ public:
     LUA_DECLARE(setElementVisibleTo);
 
     // Element data
-    LUA_DECLARE(getElementData);
+    LUA_DECLARE(GetElementData);
     LUA_DECLARE(setElementData);
     LUA_DECLARE(removeElementData);
